Add readiness probes
