+++ test/e2e/transformer/test_raw_transformer.py
@@ -1,4 +1,4 @@
-# Copyright 2019 kubeflow.org.
+# Copyright 2021 kubeflow.org.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -75,7 +75,7 @@ def test_transformer():
     except RuntimeError as e:
         raise e
 
-    time.sleep(60)
+    time.sleep(30)
 
     isvc = KFServing.get(
         service_name,
